PART 2: Change temperature type to int

+++ src/main/java/icd0004/report/ForecastWeather.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ForecastWeather {
-    private Double temperature;
+    private int temperature;
     private int humidity;
     private int pressure;
 }
